requirements.txt
tensorflow>=2.10.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
flask>=2.0.0
python-telegram-bot>=13.0
matplotlib>=3.4.0
seaborn>=0.11.0
flask-cors>=3.0.0
flask-swagger>=0.2.14
